[Discord] Rejoin voice channels on ready only if Audio cog is loaded

+++ Discord/Harmonbot.py
@@ -35,10 +35,11 @@ if __name__ == "__main__":
 		
 		if restart_channel_id := await client.db.fetchval("DELETE FROM meta.restart_channels WHERE player_text_channel_id IS NULL RETURNING channel_id"):
 			await client.send_embed(client.get_channel(restart_channel_id), ":thumbsup::skin-tone-2: Restarted")
-		for record in await client.db.fetch("DELETE FROM meta.restart_channels RETURNING *"):
-			if text_channel := client.get_channel(record["player_text_channel_id"]):
-				client.cogs["Audio"].players[text_channel.guild.id] = audio_player.AudioPlayer(client, text_channel)
-				await client.get_channel(record["channel_id"]).connect()
+		if audio_cog := client.get_cog("Audio"):
+			for record in await client.db.fetch("DELETE FROM meta.restart_channels RETURNING *"):
+				if text_channel := client.get_channel(record["player_text_channel_id"]):
+					audio_cog.players[text_channel.guild.id] = audio_player.AudioPlayer(client, text_channel)
+					await client.get_channel(record["channel_id"]).connect()
 		
 		# TODO: DM if joined new server
 		# TODO: DM if left server
